+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-3985 Added a bar chart, that shows storage usage for a space.
 * VFS-3571 Common libraries update
 * VFS-3955 Better truncating of too long names in sidebar, sidebar buttons
 * VFS-3619 Refactor of login page components